+++ devtools/pip-tools/check_sources/plugins/cmake_format_plugin.py
@@ -1,6 +1,6 @@
 #   BAREOS® - Backup Archiving REcovery Open Sourced
 #
-#   Copyright (C) 2020-2020 Bareos GmbH & Co. KG
+#   Copyright (C) 2020-2025 Bareos GmbH & Co. KG
 #
 #   This program is Free Software; you can redistribute it and/or
 #   modify it under the terms of version three of the GNU Affero General Public
@@ -17,11 +17,12 @@
 #   Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
 #   02110-1301, USA.
 
-import logging
-from pathlib import Path
+"""reformat CMake code using cmake-format"""
 
+import logging
 from cmakelang.format.__main__ import get_config, process_file
 from cmakelang.configuration import Configuration
+from ..registry import register_modifier
 
 
 def format_cmake_file(file_path, file_content):
@@ -33,25 +34,13 @@ def format_cmake_file(file_path, file_content):
     # check if formatting was aborted
     if not reflow_valid:
         logging.info(
-            "{}: contains long lines that cannot be split automatically".format(
-                file_path
-            )
+            "%s: contains long lines that cannot be split automatically", file_path
         )
 
     return out_text
 
 
-if __name__ == "__main__":
-    import sys
-
-    logging.getLogger().setLevel(logging.DEBUG)
-    print(format_cmake_file(Path(sys.argv[1])))
-    exit(0)
-
-
-from ..registry import register_modifier
-
-
 @register_modifier("CMakeLists.txt", "*.cmake", name="cmake format")
 def modify_cmake_format(file_path, file_content, **kwargs):
+    del kwargs
     return format_cmake_file(file_path, file_content)